Moved BoneTree bits to its own Interface Component class
